Adding BDD Code

+++ features/environment.py
@@ -3,5 +3,6 @@ from utilities.configurations import getHeaders
 
 
 def after_scenario(context, scenario):
-    delete_response = requests.delete(context.url + "/" + context.uid, headers=getHeaders())
-    print(delete_response.status_code)
+    if "collection" in scenario.tags:
+        delete_response = requests.delete(context.url + "/" + context.uid, headers=getHeaders())
+        print(delete_response.status_code)
